src/main/kotlin/de/bund/digitalservice/ris/norms/application/EditNormFrameService.kt
package de.bund.digitalservice.ris.norms.application

import de.bund.digitalservice.ris.norms.application.port.input.EditNormFrameUseCase
import de.bund.digitalservice.ris.norms.application.port.output.EditNormOutputPort
import de.bund.digitalservice.ris.norms.domain.entity.Norm
import org.springframework.stereotype.Service
import reactor.core.publisher.Mono

@Service
class EditNormFrameService(
    private val editNormOutputPort: EditNormOutputPort
) : EditNormFrameUseCase {

    override fun editNormFrame(command: EditNormFrameUseCase.Command): Mono<Boolean> {
        return editNormOutputPort.editNorm(Norm(command.guid, command.longTitle))
    }
}
